Dispatch/dispatch/requirements.txt
Flask
mysql-connector
pika
